#254 rename to and implement TokenProvider
